Added Debugger release mode to quickly get debug-able builds out.
Some customized build settings that help me quickly get a debuggable build up and running (with breakpoints and stuff).  With CR I would waste quite a lot of time getting this setup on each branch I worked on.  I hope this doesn't step on anyone's toes.  I'm not overly familiar with the different settings.  Decided to make this a separate commit so that it might be easier to chop out later if need be.