Fixed NullReferenceException when non-player pawns look for weapon.
This was similar to what I did in Combat Realism.  Basically when CE was looking for a gun with ammo nearby CE had assumed that all weapons in the list were ammo users (ie ranged).  By limiting the list of weapons considered to ranged weapon the NRE is avoided.

Also added a failover so that if a pawn cannot find a ranged weapon (with ammo) they will grab a melee weapon, assuming it met their other criterion besides being ranged.

It looks like the issue that popped up after this in CR didn't in CE :) which is that prisoners would get stuck in an infinite loop trying to grab a weapon they couldn't path to and could starve.