Fixed error when dropping minified thing from gear tab.
When a pawn has an empty loadout they can carry anything in their inventory.  If a pawn picks up a minified thing and then later the player hits the drop button on that thing, the error which no longer comes up was:
Verse.Pawn_InventoryTracker container tried to drop <thing> which it didn't contain.

The issue was that the gear tab shows the non-minified thing (which is probably good) but drop tried to act on that non-minified thing.

The fix was to check if the thing being dropped could be minified, if it can then search the current inventory for an item which contains the thing player tried to drop.  Used Thing.ThingID to identify.