Fix for warden stuck in a loop giving/taking prisoner food... If the food is in the warden's loadout and they want it to fulfill their loadout with it.
Tried using a predicate this time, hopefully avoided logic flaws and increasing readability from when I tried to do this in CR.

Basically (the goal is) if the pawn is of player faction, the item is nutritional, and the item is in a prisoner then the thing is ignored by a pawn trying to fulfill their loadout.  Reasoning is that you can't haul food from a prisoner's room.

Didn't add "using System;" at the top of the file for just one additional item...