Enhancement: Weapons a player can't get aren't shown for loadouts.
This is a pretty minor and almost cosmetic alteration (also one I've done in CR).  The idea is if a weapon is one that a player's pawn can't get their hands on, don't show it in the loadout configuration.

I used menuHidden property as the deciding factor but I suppose dropOnDeath would work too.  Generally speaking menuHidden weapons are those which are on turrets, in the hands of mechanoids, or otherwise not meant for player's pawn's use.

I decided to toss this check onto melee weapons as well, though the situations where this is apt seems REALLY uncommon.