Enhancement to make loadout slots unique (per def).
I didn't really test this in CE (just realized) but in CR things could get weird if you had 2 slots tracking the same thingdef (ie 2 pistols in loadout).

The idea is when a new slot is added to the pawn's loadout a check is done to see if that thing (def) is already being tracked by their loadout.  If it is then the existing loadout has it's count incremented, if not then the new slot is added to the loadout as usual.