Feature: Add minifiable things to loadouts.
With the addition of caravans and especially with certain mods that add new camping gear it has become somewhat desirable (to me) to include minified things into a loadout.

This touches bits that a lot of things use (loadout system) and when I added this to CR it caused a fair number of problems which I think I managed to find pre-merge.  I'm still not the most comfortable with C#...