Modified DrawnWornExtras to reduce branching and assignments.
Some of the changes include the fact that wearer.Spawned returns false if a pawn is being carried or is dead so was able to chop some parts out there.

Some additional code to handle moving some features into xml and expand capability towards Apparel_VisibleAccessory on the head.

The xml part hasn't been done as of this time and the mod is in a position to have that attempt and added feature chopped out easily while retaining the new logic.