Heavily modified Apparel_VisibleAccessory.
Functionally it should work the same where visible accessories are above the shell layer but below head layer in all situations where it would draw such an accessory (when a pawn is not dead and not being carried for example).

Tried to reduce assignments (since I don't know how expensive some were) as well as branches in the code while retaining the logic that was present before.

There are now several constants as well as updated reference relating to different layers to aid in maintenance without having to dig through decompiled code. Proper altitude layers are consulted for different states of the pawn.

New workings - Apparel_VisibleAccessory is no longer abstract class, instead a new xml field "order" has been defined with a range of 1 to 4.  The upper bound may change pending further tests.  Apparel_VisibleAccessoryDef class handles the order and a helper function to clamp the value.  If order goes unclamped things could get weird looking.

'removed' Apparel_Backpack and Apparel_TacVest classes.  They aren't deleted but just disabled from being used in compile.

GetAltitudeOffset method functions similar to the old non abstract extended classes though now that various offsets are known it was less useful to extend the class.  This makes use of the order in the def and decides if a particular offset should be used.

New functionality (currently disabled pending testing/development) to add VisibleAccessory headgear.  Avoid using until this code path can be tested/developed further.

New helper property: onHead uses a cached dictionary of results.  How it works is if a value is not cached for a given def then for every body part that the apparel covers the bodypart parent is inspected until either the torso or head is found.  From that we know if the apparel is on body or on head and thus which offset to use.