Projectile physics refactor + minor bugfixes
- ProjectileCE.Launch() now takes shotAngle, shotHeight, shotSpeed and
the new shotRotation (which is an angle in degrees between the shooter
and the target).
- An additional parameter for minimal collision distance was added for
the time being.
- Explosions can now spawn at different heights and fragments are
ejected in a sphere around the center.

Detours_PawnEquipmentTracker now checks whether primaryInt is null
before checking whether PrimaryEq is null - it was possible to start an
attack twice with grenades and cause an error when the second warmup
finished and no grenades were left to throw.