Issue 46 take two (#92)
* Partial Issue 34 - Nothing loadout cannot be edited.
This was low hanging fruit for me and helped me get to know a little more about the loadout system as a whole.

I make the assumption that the default flag is only on the Nothing loadout.  So from there just don't show the Nothing loadout in the edit interface.

Additionally I had to make some changes to the AI system to compile.

* Added Loadout Copy feature.
So now when a player is in the loadout configuration screen and they are on a loadout they can hit copy and have a new loadout generated based on the current loadout.

Modified Dialog_ManageLoadouts to support the new button (and the 2 lines of code that make it work).

Modified MapComp_LoadoutManager to add a method overload to generate a unique label from a given string.

Modified Loadout and LoadoutSlot to support a Copy() method and a static Copy(object) method to handle making copies.

* Fixed animal inventory drop issue.

* Added ability for pawns to fulfill loadout needs from animals.
Also from prisoners...

Modified JobGiver_UpdateLoadout to look inside inventories of other pawns on the same map as the pawn with the loadout missing items once the other two item searches fail.

Added a new JobDriver TakeFromOther which handles the task of getting the pawn to the pawn with the item they want, reserving it, a brief pause and then taking the item.

If the item being taken is flagged to be equipped then uses CE's code to swap the item from inventory to equipped.

* Did a little touch up and testing...
Tried to limit how many pawn inventories are checked.
Also tried to use Pawn.inventory.GetInnerContainer() method instead of Pawn.inventory.innerContainer field following a suggestion that the field was probably intended to be private.

My knowledge of how things get optimized in .net is still very basic so not sure if what I did to limit the inventories checked helps or if it ends up neutral.

Basically after both searches of the map come up empty for an item it will now locate all pawns on the same map which are either pack animals of the same faction as the searching pawn or prisoners hosted by the same faction as the searching pawn... as long as their inventories are not empty.  It will then try to fetch an item which matches the desired def much like if it found a hit before.