All CEVerb weapons that use ammo should do a reload before warmup. (#109)
* All CEVerb weapons that use ammo should do a reload before warmup.

Took some time to find the interesting point where a pawn switches from trying to shoot to the warmup but finally found it.  From there did a Harmony Hybrid Infixed Prefix (functions like a Prefix but is Infixed).

The infix patch passes the instance verb into a needs reload check.  If the check returns false that means there was work to do (yeah, backwards) and the pawn shouldn't enter the warmup stance.  If true then either it wasn't a gun with ammo or it didn't need reload, the pawn enters the reload stance.

Thanks to whoever created/maintained the CompAmmoUser.TryStartReload() as I had no idea how to get turrets reloading and that just handled it for me.

Also includes some side fixes related to loadouts.

* Forgot to turn off Harmony Debug again...