Loadouts auto reload job+ (#114)
* Implemented Loadouts sorting.

Default is always first.
All other loadouts are sorted by name string.
Loadouts are sorted before showing the user a list of loadouts.
The base loadouts list in LoadoutManager class is sorted so original creation order is lost.  Can be modified to return a new sorted list instead...

* Added some optimizations to ManageLoadouts UI.

Specifically found an optimization I forgot about for Generic Ammo so the predicate can fail out sooner.

Also spread generic testing (if there is an item on the map satisfying the generic) out over time so that in any given frame only one generic is being tested.

Also modified Generic and Specific testing to instead of finding EVERY item on the map which satisfies it it will stop on the FIRST item that is satisfied so performance should be more consistent on very busy maps rather than slowing down.

* Caused the 'Add ammo for ...' UI to give a clip's worth of ammo.

So for example a pistol which holds 7 ammo, using the float menu instead of the list will give 7 (more) ammo.

Also should ammo default counts ever be implemented, it will fall back to that (useful for arrows and the like where there isn't a clip).

* Initial addition of automatic reload after UpdateLoadout.

Pawns are now concerned about 2 things:
1 - Are any of their guns low on ammo?
2 - Are any of their guns loaded with ammo that they don't have a LoadoutSlot/HoldRecord for a clip's worth?

If either of those are true and there is enough (approved by LoadoutSlot/HoldRecord) ammo, the pawn will do a reload (and a switch ammo if necessary).

* Fixed a bug in AttackTargetFinder patch code.

* +Generate Loadout from Inventory & MoteThrower to JobDriver_Reload

New button (suggestion of button from Karim) on Pawn's gear tab which creates a new Loadout from that Pawn's inventory.  Button only shown if the Pawn has no/empty Loadout and has equipment/inventory items.

Possible refinement: Only show the button if the items in inventory are not marked by Generics that are DropExcess only.

Also moved the motethrower associated with reloading into the JobDriver and out of two other sections of code.  May throw an exception if the creator of the job did insufficient error checking.

* 3 Bug Fixes...

-A new bug due to my lack of understanding how ThingContainer.TryAdd() works caused an attempt to destroy a thing twice (basically ended up trying to add a thing to a container AND drop it on the ground).

-Fixed a bug introduced when the default loadout became non-editable.  Basically an NRE.

-Fixed a really old bug (from before I got my hands on the code) involving creating an IEnumerable from a Dictionary and then altering the Dictionary (in particular changing things so that the IEnumerable clauses were no longer valid).  When I started on this code base the Remove button in Manage Loadouts wasn't properly removing Loadouts via the LoadoutManager but rather just from the list.  Early on I changed that to what seemed more correct and exposed the bug.  Since the list being manipulated by Remove button was the same object stored by LoadoutManager it worked but pawns with that Loadout assigned weren't having their assignment changed (along with anything else LoadoutManager might have done).

* Further separated out functions of CompAmmoUser.TryStartReload()

The Unload mechanic has been separated off (CompAmmoUser.Unload()) so that it can be called on it's own though TryStartReload() still takes a bool indicating to only perform an unload.  More work necessary to finish separating that off.  Though TryStartReload does error checking that Unload() does not.

Also Unload() will drop any items on the ground that didn't fit into inventory instead of destroying them.

Added CompAmmoUser.GetReloadJob() which takes a bool indicating if the job should be flagged PlayerForced or not.  This doesn't do any error checking and expects the caller to check for errors (atm).

This also sets the storedTarget and storedJobDef to null so the caller doesn't have to do that, if they should have some object stored in them the caller can do that after getting the Job.

More to do...

* Some more tweaks on CompAmmoUser.TryStartReload.

Changed some logic around to clean things up a little.

Renamed Unload() to TryUnload().  Returns false if there is something wrong with the weapon (would indicate can't do the reload job).  Returns true if the weapon is already unloaded or if the unload was successful.

Added an warning message if the ammo was destroyed because it couldn't be dropped.

Changed GiveJob() to TryGiveJob() and modified it's logic as well.

Modified TryStartReload() so that it no longer takes a bool.  Instead those wishing to unload a weapon would call TryUnload().  Modified all in project found references to CompAmmoUser.TryStartReload(true) to instead use CompAmmoUser.TryUnload().

Testing suggests behavior has been maintained.  Turrets and Pawns, if clip is empty, initiate a reload before aiming/warmup and resume the tasks they were doing before.  Pawns will initiate a reload when not otherwise busy when their clip is low on ammo.

Untested: Ammo in a gun not equipped gets reloaded.

* Attempt to have JobGiver_CheckReload's DoReloadCheck use cache.

Cache provided by CompInventory.

Still does a fair bit of enumeration but the cache already excludes a lot of stuff so the enumeration count is smaller.

* Cleanup of remarks from mid-development

* Bug Fix + Inventory Reload guns and Reloading mote on turrets.

First, first a bug in Utility_HoldTracker, GetStorageByThingDef could try to double add a dictionary key if ammo for a weapon had already been logged in the open and more ammo was found for a given gun.  Proper key check is in place now.

Next tested attempts to reload guns in inventory (this was a requested feature iirc) and found that the code was VERY specifically built to ONLY reload equipment and not inventory items.

Added a new property to CompAmmoUser named holder which is either the existing wielder property (equipment) or the pawn's inventory that is storing the item.  Should return null if the container is not a pawn's inventory at which time can't do a reload.

Additionally JobDriver_Reload was modified to use CompAmmoUser.holder instead of CompAmmoUser.wielder.  Code paths off of that (ie CompAmmoUser.LoadAmmo()) also use holder instead of wielder now.  wielder has a very specific sense that I didn't want to break in other code that I didn't look at.

Fixed a bug (old?) in JobDriver_Reload related to waitToil not having it's actor filled and thus the StopDead() clause never triggered properly.

Modified HasNoGunOrAmmo condition check (FailOn) so that if an item started out in equipment and was moved during reload, the reload is interrupted.  Same for if an item was in the inventory but is no longer.  Additionally a check for if a pawn's equipment changes at all to interrupt the reload.

Modified the "Reloading!" Mote (now that I understand what that was doing in it's original context) to be pawn specific and doesn't care about turrets.

Added the mote throwing code from 'original' CompAmmoUser.TryStartReload() related to turrets into JobDriver_ReloadTurret.

I didn't encounter any big bugs in my narrow test of firing off a turret (pawn manned) while another pawn auto-reloads their Survival Rifle (equipment) and then Pistol (Inventory).