Pawn auto firing issues (#124)
* Fixed issue #120

Or at least hopefully fixed #120 regarding the red error text.  Not sure about the other issues as I haven't been able to replicate.  Pawns with multiple weapons and ammo definitions seem to pick things up fine.

The issue seemed to be caused by a failure to check if the weapons in the inventory rangedWeaponCache had CompAmmoUser before trying to use said comp.

* Attempt to fix various issues related to drafted pawns auto-firing.

Issue 1: The code that inserted the call in IL was inserting it in the wrong place, breaking the key logic of the function.

Issue 2: The call to Verb_LaunchProjectileCE.GetPartialCoverBetween() had the 2nd and 3rd arguments backwards from the correct order, so was checking if the shooter had cover and if so the target was invalid (heh).

Issue 3: An older bug in GetPartialCoverBetween where the checkedCells list wasn't being updated with the cells which had been checked potentially reducing performance in all situations related to shooting.