Dev loadouts to game comp (#206)
* Converted LoadoutManager class to GameComponent.

This involved making most of the class instance based rather than static.  Also had to learn a few things about how GameComponent works.

This commit has HoldTracker data not saved/loaded since I'm stuck on an odd problem with that.  Is save compatible in that you can load your save game but any Loadouts (and HoldTracker data) in the save will be dropped.

* Very little progress...

Committing what I've got (slightly more functional) so I can look at something else.

* Modified HoldRecords to more closely follow Loadouts.

This fixed the save/load issue I was having.  There might have been another solution but I knew this path would work (since Loadouts worked) so I didn't pursue that other path.

Also this commit returns LoadoutManager to a mostly static interface.

Since there is a window where LoadoutManager should not be returning/accepting actual data there are checks in place to see if it's apt to return something and if not, doesn't return anything.

This is because there is a risk of data bleeding from one instance to the next when loading a save game while an active game is running.