Modified JobGiver_UpdateLoadout and JobDriver_TakeFromOther... (#247)
* Modified JobGiver_UpdateLoadout and JobDriver_TakeFromOther...
In order to prevent an error which could occur when multiple pawns tried to take something from a pack animal...

Also modified JobGiver to check for reservation and reachability.

Modified JobDriver to make sure that the Thing the pawn wants to take is still in it's inventory and if the stack fell below what was desired take the rest of the stack.

Also added some additional fail conditions that seemed logical.  One might be redundant...

NOTES:
 1 - Need to test if my read of ReservationManager.CanReserve is accurate and if not either set reservations down to 1 (quick change in two sections of code) or add a more detailed check for any jobs reserving the animal are NOT JobDriver_TakeFromOther as a block.
 2 - The check if the thing is still in the takeFrom pawn's inventory isn't cheap however it's necessary as if a pawn took a full stack of something from the animal the Thing object was transferred rather than a new thing created meaning that it wouldn't be in takePawn's inventory anymore.

* Modified the thing check to be faster...
Now the Thing Destroyed check isn't redundant...
Instead of searching the takePawn's inventory (could be 10+ items) it does a check for if the pawn holding the takeThing is the takePawn's inventory.
This basically walks a linked list from Thing to Pawn...  Should generally (on average) be faster.