Fix for issue #260
Implemented prefix to JobGiver_DropUnusedInventory. The forced drop was screwing up loadouts so now it will check if the pawn in question has a loadout before dropping. For raw food being dropped it will clear any forced hold records. In the case of drugs it will do nothing (since the vanilla functionality seems to make a lot of assumptions about whether a pawn should be allowed to carry a drug he is currently high on which we replace with direct player control).