Projectile arcs
- Height is now displayed visually on ALL projectiles (including
fragments which now take lots of computational power).
- Projectiles shot by a "launcher != null" are only displayed after
their first tick (so they don't spawn in the middle of the pawn).
- Projectile rotation follows the projectile arc.
- A correction factor for all arcs exists such that the projectile
visually approaches height==0 as it reaches the end of the trajectory
while it starts off with a correction for the shotHeight of the
projectile so it matches the vanilla projectile position.
- DrawPosition does never match ExactPosition for projectiles anymore.