Attempted to rebalance Vanilla Armour Expanded medieval armor.
Prior to this patch, it's be possible to have 0.813 sharp armor rating (steel light plate armor, chestplate and chainmail) on the torso (for reference power armor is at 0.8) for the mass of 25. after this patch, it's only possible to have 0.488 sharp armor rating (steel plate armor and chainmail) on the torso (FMJ .44 can't pierce, but FMJ .303 British and AP .44 can 100% of the time) for the mass of 20.