Explicitly handle case where worn_apparel ends up null.
The dotnet compiler is happy with `foreach (var a in maybenull?.something)`, but I can't find any documentation on how it's implemented internally (or even that it's officially supported as of C# version 8).
The mono compiler chokes if given that input.  Moving the iterable to its own line silences the compiler error, but I think it might then try to iterate over null, which would then raise a runtime exception.  
This uses the syntax that's been supported for a long time, so should work reliably into the future and with any C# compiler.  
As for performance, the JIT *should* know that Enumerable.Empty is always empty and skip the for loop, so the performance *should* be the same.