update the scorcher helmet
The 31.07 feral faction update changed the stats of the Scorcher Helmet, removing the blunt and sharp armor rating stats. Combat Extended is trying to replace these attributes with new values, but they no longer exist and thus the error is thrown. As such, the CE blunt and sharp armor stats are added