[VAE] Buff otter stats
Currently, VAE Boreal Forest's Otter is rather weak and prone to getting
mauled and even killed by small prey animals (rats, hares, squirrels
etc.). This patch aims to buff the otter's stats to make them more
realistic for a predator of its kind and size. I've opted to bring the
otter's melee stats closer to those of the RimWorld fox, as both the Eurasian
otter[1] and the North American otter[2] are similar in terms of mass
and body size to the red fox[3].

Specifically:
* Increase melee dodge chance to 24%, up from 5% (it's 27% for foxes)
* Increase melee critical chance to 2%, up from 0 (2% for foxes)
* Add surprise attack effect to the 2x claws and 1x teeth attack
  providing stun and a bonus 20 damage (just like foxes)
* Reduce cooldown time of claws from 3.7 to 1 (0.91 for foxes)
* Reduce cooldown time of teeth from 3.3 to 1.5 (just like foxes)
* Increase armor sharp pen for teeth to 1.5, up from 1.1 (just like
  foxes)

---
[1] https://en.wikipedia.org/wiki/Eurasian_otter
[2] https://en.wikipedia.org/wiki/North_American_river_otter
[3] https://en.wikipedia.org/wiki/Red_fox