Prevent using ranged weapons with shield belts
In #767, the old shield belt patch was disabled due to breakage, which
now allows for ranged weapons to be used even if their wielder is
wearing a shield belt. Fix the shield belt patch to correctly consider
both vanilla and CE ranged verbs when determining if they can be used
while wearing a shield, maintaining the existing exception for CE
binoculars.

Also remove various dead code from the shield belt patch. Vanilla shield
belts no longer run a reachability check in 1.3 and the relevant part of
the patch got commented out in #767, so let's get rid of it. Likewise,
the turret-related parts also seem to be dead code, as the
TurretsBreakShields settings only controls whether a pawn's shield belt
should be disabled while they're manning a turret, not whether they can
fire the turret - which they already can do irrespective of this patch.