Avoid potential loop while reloading turrets
Currently, the jobgiver for reloading turrets may give out a job to
reload a turret with a stack of ammo that's currently cooking off, which
results in a loop as it causes the jobdriver to fail the job. As a fix,
ensure that the jobgiver does not attempt to use ammo which is cooking
off.

For correctness, also make sure that ammo types which can never cook off
(e.g. 12x64mm mechanoid ammo) are never considered to be cooking off.