Reduce fire starting chance of fire arrows
Make sure that fire arrows don't start a fire on every hit. Flame damage
type effectively guarantees that the target will be set on fire
irrespective of explosionChanceToStartFire, and CE chemfuel puddles also
autoignite, so use Arrow damage type instead with a chance to spawn a
chemfuel puddle (that will then ignite).

The chance to spawn chemfuel is kind of a placeholder value atm. We
could fine-tune it with additional testing.