npc targetting and internal organ attacks
npcs now target torso if it's not armored and head if it's armored. It might sound odd, but hitting internal organs is fixed now, so torso hits will be much more lethal also considering that it's easier to hit liver/kidney/stomach etc. than to hit brain on head hit