Small fix for Horse Mask
From Aquiles#4265 on discord:
After many hours of testing, I found the error in the patch for Vanilla Ideology Expanded - Hats and Rags. The CE patch seems OK at first glance, but to make the error disappear a small change is needed: in the Apparel_Headgear file, the PatchOperation for the HorseMask should be Replace, not Add. Just in case you want to merge it to the github version.