Add Shatterer
I am not sure how to balance this one. I upscaled Harvester stats a bit as it's also a scythe of sorts and added Royalty's Warhammer stats for blunt attack and upscaled it too. Offsets are also from Harvester and -8 melee offset is preserved from vanilla variant of Shatterer. 

In other words, it works for me but can probably use some balancing work.