Make VerbProperties.LaunchesProjectile cache vary per property
123cc55e898a17c3baedb56868586c3f9eb8fe3f added caching to our Harmony
patch that uses reflection to determine whether a given verb is melee
or not, but the cache only varied on the type of the VerbProperties class,
which is always VerbPropertiesCE. This could result in e.g. melee verbs
getting classified as ranged ones. Let's fix it by varying the cache
per property instance.