LetterTracker: Fix invalid Mech_Centipede def reference
With 1.4, Mech_Centipede is now an abstract def that is used as a base
for different variants, like Mech_CentipedeBlaster and
Mech_CentipedeGunner (minigun). Accordingly, LetterTracker throws an
error when it tries to use this abstract def to compute the armor value
of centipedes. Let's use the Mech_CentipedeBlaster def instead which
should be suitable for our purposes here.