Fix black smoke rendering and propagation
Black smoke uses the pre-1.4 object gas system, which is still present
in core, but the BaseGas def has been removed along with some fields
from GasProperties. Augment the smoke def with missing data that used to
be in the base def, and introduce a new CE-specific gas property class
to store the fields removed from core GasProperties.

This does not yet fix the interaction of combat with black smoke - we
should do that in a followup patch.