Fix drafted pawns becoming guilty when drafted & ordered to attack
Ensure pawns won't be considered guilty when drafted and ordered to
attack a friendly pawn. This prevents an exploit where players
could draft an undesired pawn and have them punch a colonist or colony
animal so that they'd become guilty & be banished with a lesser impact
on the colony.

Both melee and ranged attacks are covered.