Fix new Centurion/Legionary shields
Biotech's new Centurion/Legionary mechs use shields with finite
hitpoints that break if they take enough damage. Our custom intercept
logic in ProjectileCE must be updated to handle these shields.

Solve this by just applying the appropriate vanilla logic - applying
damage from the projectile to the shield and breaking it if it reaches
zero hitpoints. For this to work, move the definition of DamageAmount
from BulletCE to ProjectileCE so that we can obtain the scaled damage
amount in the interception logic.