Optimize running from cover from suppression
Only try to create a new "run for cover" job if the pawn is not
currently running for cover. Currently, we create it every time when a
pawn is suppressed and is above the threshold, but discard it anyways if
they're already running for cover.

When profiling the immediate period of a doomsday explosion in a raid,
this reduced the impact of AddSuppression() from ~40% to ~1.3.%.