Prevent busted mod assemblies from breaking CE on load
If the active mod list contains mod assemblies incompatible with the
game (i.e. they throw a ReflectionTypeLoadException when their types are
fetched), vanilla's ModAssemblyHandler logs an error and skips over
those assemblies. However, we try to process all loaded assemblies as
part of the initialization stage where we're looking for compatibility
patches, so if such a broken mod is present in the modlist, it will
break CE initialization and lead to various NREs in-game.

As a fix, harden the loading process to catch type load errors from
assemblies. Since these errors are likely to be from some unrelated,
broken mod, only log a message if development mode is active to reduce
user confusion and logspam.