Reduce the performance impact of GetHighestCoverAndSmokeForTarget()
GetHighestCoverAndSmokeForTarget() needs to iterate over all cells on
the line of sight to the target and also needs the count of cells on the
line. It currently does this by converting the cells IEnumerable to an
array via ToArray(), which always allocates a new array and can result
in considerable memory pressure when many pawns are in combat and
looking for targets. Reduce the memory churn by introducing a new helper
method for getting a list of cells on a line of sight that uses a
preallocated List under the hood to store the cells.

We can further reduce the impact of cover checking in the subsequent
iteration by only calculating cover height once we've determined that
the cover is not a plant, as the calculation does not consider plants
anyways in this particular case.