Remove the simplified ammo system
Simplified ammo is a common trap for people that want to avoid
micromanaging many ammo types but end up finding themselves defenseless
against threats that could be easily dealt with using advanced ammo
types. The generic ammo system is a superior alternative for reducing
the number of ammo types to manage, so there's no need for the
simplified ammo toggle to exist now.