Rimsenal Assault Armor changes
Modified the movespeed penalty on the helmet slightly. Increased the provided carry weight for the body armor itself.
This should result in an overall simmilar movespeed penalty from the armor from the current configuration, without making the armor exceptionally good when moving capacity is scaled up. Additionally leaves the armor as heavily armored but still somewhat less optimal that cataphract given the price difference in the two.