Nerfed Hyperweave Heat Protection to 2x Devilstrand
Hyperweave has less than 2x the mass multiplier of Devilstrand, and the 1.92 heat armor multiplier is making high quality versions of normal clothes (such as tribalwear or jumpsuits) have over 800%+ heat resist. The heat armor multiplier on Devilstrand is 0.05. This adjustment nerfs Hyperweave to 0.1, so it is still twice as strong as devilstrand.