Fix retaliation raids and shelling
As reported in #3419, retaliation raids currently cause infinite error spam
if the bombardment that triggered the retaliation was from a temporary map
that was disposed of by the time the raid was actually executed. Fix it by
checking whether the target map for the raid is actually valid.

While testing this, I found that retaliation shelling has also been broken since d90f3e1395438165e9bade2780299cb2b64b177d.
Fix it by accounting for the case where the launcher is not spawned at all,
since retaliation bombardment shells are "launched" by world pawns.
Also ensure bombardments from other factions' settlements do not trigger a retaliation
bombardment if the target faction isn't hostile to the owner of the tile,
and introduce a configurable retaliation fire delay that scales with how damaged a
site is.